Aces-Full house options completed
finished the score calculations for aces, 2s, 3s, 4s, 5s, 6s, three of a kind, four of a kind, full house, and yahtzee.
completed dice and yahtzee class, along with a testHarness class with a main to test everything
in yahtzee, still have to finish calculations for some of the scores (include private helper methods for those) and also the score() method. almost done